Added all necessary binding

+++ app/src/main/java/com/example/android_task_4_2/GameFragment.kt
@@ -90,7 +90,7 @@ class GameFragment : Fragment() {
 
         mainScope.launch {
             val gameResult = GameResult(
-                timestamp = System.currentTimeMillis().toInt(),
+                timestamp = System.currentTimeMillis(),
                 computer = computer,
                 you = option,
                 result = result
